fix: move key generation / management to hook utility
