+++ src/Host.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { View, ViewStyle } from 'react-native';
-
+import { useGeneratedKey, removeKey } from './hooks/useGeneratedKey';
 import { Manager, IManagerHandles } from './Manager';
 
 interface IHostProps {
@@ -44,36 +44,8 @@ export const Host = ({ children, style }: IHostProps): JSX.Element => {
     }
   }, []);
 
-  //Keep track of all the keys we've generated to ensure uniqueness
-  let usedKeys: Array<string> = [];
-
-  //Makes a magical random key
-  const keyGenerator = (): string => {
-    return `portalize_${Math.random().toString(36).substr(2, 16)}-${Math.random()
-      .toString(36)
-      .substr(2, 16)}-${Math.random().toString(36).substr(2, 16)}`;
-  };
-
-  //Keeps making new keys until a unique one has been found
-  const getNewKey = (): string => {
-    let foundUniqueKey = false;
-    let newKey = '';
-    let tries = 0;
-    while (!foundUniqueKey && tries < 5) {
-      //limit number of tries to stop endless loop of pain
-      tries++;
-      newKey = keyGenerator();
-      if (!usedKeys.includes(newKey)) {
-        foundUniqueKey = true;
-        usedKeys.push(newKey);
-      }
-    }
-
-    return newKey;
-  };
-
   const mount = (children: React.ReactNode): string => {
-    const key = getNewKey();
+    const key = useGeneratedKey();
 
     if (managerRef.current) {
       managerRef.current.mount(key, children);
@@ -104,7 +76,7 @@ export const Host = ({ children, style }: IHostProps): JSX.Element => {
   const unmount = (key: string): void => {
     if (managerRef.current) {
       managerRef.current.unmount(key);
-      usedKeys = usedKeys.filter(k => k !== key); //remove our key to make it 'available' again
+      removeKey(key);
     } else {
       queue.push({ type: 'unmount', key });
     }
